Add AccountServer mailbox queue limit
With this before sending any message to the AccountServer the code will
check if it's queue is already full. If it is the message will be
dropped with {:error, :process_mailbox_is_full}.

I've opted to use the Process.info to check the queue size, even though
we should avoid using it in production code.

That is because in real life a hard limit on messages would probably
not be the best way to tackle this issue (a timeout for queue time
for example would be a lot more flexible and better accomodate
functions with different execution times).

Another option I've considered was building the queue code myself, but
it would increase the AccountServer complexity by a fair bit and I
thought the Process.info approach should be enough for the exercise.